fix(components): [scrollbar] multiple rendering
