+++ packages/components/scrollbar/src/bar.ts
@@ -7,16 +7,6 @@ export const barProps = buildProps({
     type: Boolean,
     default: true,
   },
-  width: String,
-  height: String,
-  ratioX: {
-    type: Number,
-    default: 1,
-  },
-  ratioY: {
-    type: Number,
-    default: 1,
-  },
 } as const)
 export type BarProps = ExtractPropTypes<typeof barProps>
