+++ packages/components/scrollbar/src/constants.ts
@@ -3,6 +3,7 @@ import type { InjectionKey } from 'vue'
 export interface ScrollbarContext {
   scrollbarElement: HTMLDivElement
   wrapElement: HTMLDivElement
+  minSize: number
 }
 
 export const scrollbarContextKey: InjectionKey<ScrollbarContext> = Symbol(
