+++ packages/components/scrollbar/src/scrollbar.vue
@@ -20,14 +20,7 @@
       </component>
     </div>
     <template v-if="!native">
-      <bar
-        ref="barRef"
-        :height="sizeHeight"
-        :width="sizeWidth"
-        :always="always"
-        :ratio-x="ratioX"
-        :ratio-y="ratioY"
-      />
+      <bar ref="barRef" :always="always" />
     </template>
   </div>
 </template>
@@ -45,7 +38,6 @@ import {
 import { useEventListener, useResizeObserver } from '@vueuse/core'
 import { addUnit, debugWarn, isNumber, isObject } from '@element-plus/utils'
 import { useNamespace } from '@element-plus/hooks'
-import { GAP } from './util'
 import Bar from './bar.vue'
 import { scrollbarContextKey } from './constants'
 import { scrollbarEmits, scrollbarProps } from './scrollbar'
@@ -69,12 +61,7 @@ let stopResizeListener: (() => void) | undefined = undefined
 const scrollbarRef = ref<HTMLDivElement>()
 const wrapRef = ref<HTMLDivElement>()
 const resizeRef = ref<HTMLElement>()
-
-const sizeWidth = ref('0')
-const sizeHeight = ref('0')
 const barRef = ref<BarInstance>()
-const ratioY = ref(1)
-const ratioX = ref(1)
 
 const wrapStyle = computed<StyleValue>(() => {
   const style: CSSProperties = {}
@@ -135,26 +122,7 @@ const setScrollLeft = (value: number) => {
 }
 
 const update = () => {
-  if (!wrapRef.value) return
-  const offsetHeight = wrapRef.value.offsetHeight - GAP
-  const offsetWidth = wrapRef.value.offsetWidth - GAP
-
-  const originalHeight = offsetHeight ** 2 / wrapRef.value.scrollHeight
-  const originalWidth = offsetWidth ** 2 / wrapRef.value.scrollWidth
-  const height = Math.max(originalHeight, props.minSize)
-  const width = Math.max(originalWidth, props.minSize)
-
-  ratioY.value =
-    originalHeight /
-    (offsetHeight - originalHeight) /
-    (height / (offsetHeight - height))
-  ratioX.value =
-    originalWidth /
-    (offsetWidth - originalWidth) /
-    (width / (offsetWidth - width))
-
-  sizeHeight.value = height + GAP < offsetHeight ? `${height}px` : ''
-  sizeWidth.value = width + GAP < offsetWidth ? `${width}px` : ''
+  barRef.value?.update()
 }
 
 watch(
@@ -189,6 +157,7 @@ provide(
   reactive({
     scrollbarElement: scrollbarRef,
     wrapElement: wrapRef,
+    minSize: props.minSize,
   })
 )
